requirements.txt
fastapi
pydantic
python-dotenv
httpx
uvicorn
openai